+++ source/Messaging.Application/Transactions/Aggregations/RetreiveAggregationResult.cs
@@ -18,12 +18,11 @@ using System.Threading;
 using System.Threading.Tasks;
 using MediatR;
 using Messaging.Application.Configuration.Commands.Commands;
-using Messaging.Application.Configuration.DataAccess;
 using Messaging.Domain.Transactions;
 using Messaging.Domain.Transactions.AggregatedTimeSeries;
 using Messaging.Domain.Transactions.MoveIn;
 
-namespace Messaging.Application.Transactions.AggregatedTimeSeries;
+namespace Messaging.Application.Transactions.Aggregations;
 
 public class RetrieveAggregationResultHandler : IRequestHandler<RetrieveAggregationResult, Unit>
 {
