+++ source/Messaging.Application/Transactions/Aggregations/SendAggregatedTimeSeriesHandler.cs
@@ -23,7 +23,7 @@ using Messaging.Domain.OutgoingMessages;
 using Messaging.Domain.Transactions;
 using Messaging.Domain.Transactions.AggregatedTimeSeries;
 
-namespace Messaging.Application.Transactions.AggregatedTimeSeries;
+namespace Messaging.Application.Transactions.Aggregations;
 
 public class SendAggregatedTimeSeriesHandler : IRequestHandler<SendAggregatedTimeSeries, Unit>
 {
