+++ source/Messaging.IntegrationTests/Application/Transactions/AggregatedTimeSeries/WhenAggreationOfTotalProductionIsCompletedTests.cs
@@ -15,7 +15,7 @@
 using System.Threading.Tasks;
 using Dapper;
 using Messaging.Application.Configuration.DataAccess;
-using Messaging.Application.Transactions.AggregatedTimeSeries;
+using Messaging.Application.Transactions.Aggregations;
 using Messaging.IntegrationTests.Fixtures;
 using Xunit;
